extract context to separate submod
