+++ frozen-abi/src/stable_abi/context.rs
@@ -1,5 +1,3 @@
-use rand::RngCore;
-
 /// Context for `StableAbi<...>` impls on sequence-like collections
 /// (`Vec`, `VecDeque`, `HashMap`, `BTreeMap`) that bounds the sampled length
 /// to an inclusive range `min..=max`.
@@ -10,8 +8,8 @@ use rand::RngCore;
 /// `SequenceLenMax(max)`).
 #[derive(Clone, Copy, PartialEq, Eq, Debug)]
 pub struct SequenceLenRange {
-    min: usize,
-    max: usize,
+    pub(super) min: usize,
+    pub(super) max: usize,
 }
 
 /// Context for `StableAbi<...>` impls on sequence-like collections
@@ -21,16 +19,3 @@ pub struct SequenceLenRange {
 /// impl delegates to the `SequenceLenRange` impl.
 #[derive(Clone, Copy, Debug, PartialEq, Eq)]
 pub struct SequenceLenMax(pub usize);
-
-pub trait StableAbi<Ctx = ()>: Sized {
-    fn random_with_context(rng: &mut (impl RngCore + ?Sized), ctx: Ctx) -> Self;
-
-    fn random(rng: &mut (impl RngCore + ?Sized)) -> Self
-    where
-        Ctx: Default,
-    {
-        Self::random_with_context(rng, Ctx::default())
-    }
-}
-
-mod impls;
